+++ src/bin/firebase.ts
@@ -96,7 +96,10 @@ if (utils.envOverrides.length) {
 logger.debug("-".repeat(70));
 logger.debug();
 
+import { enableExperimentsFromCliEnvVariable } from "../experiments";
 import { fetchMOTD } from "../fetchMOTD";
+
+enableExperimentsFromCliEnvVariable();
 fetchMOTD();
 
 process.on("exit", (code) => {
@@ -160,9 +163,7 @@ if (!handlePreviewToggles(args)) {
   cmd = client.cli.parse(process.argv);
 
   // determine if there are any non-option arguments. if not, display help
-  args = args.filter((arg) => {
-    return arg.indexOf("-") < 0;
-  });
+  args = args.filter((arg) => !arg.includes("-"));
   if (!args.length) {
     client.cli.help();
   }