+++ src/emulator/downloadableEmulators.ts
@@ -27,14 +27,14 @@ const CACHE_DIR =
 
 export const DownloadDetails: { [s in DownloadableEmulators]: EmulatorDownloadDetails } = {
   database: {
-    downloadPath: path.join(CACHE_DIR, "firebase-database-emulator-v4.9.0.jar"),
-    version: "4.9.0",
+    downloadPath: path.join(CACHE_DIR, "firebase-database-emulator-v4.10.0.jar"),
+    version: "4.10.0",
     opts: {
       cacheDir: CACHE_DIR,
       remoteUrl:
-        "https://storage.googleapis.com/firebase-preview-drop/emulator/firebase-database-emulator-v4.9.0.jar",
-      expectedSize: 34204485,
-      expectedChecksum: "1c3f5974f0ee5559ebf27b56f2e62108",
+        "https://storage.googleapis.com/firebase-preview-drop/emulator/firebase-database-emulator-v4.10.0.jar",
+      expectedSize: 34230230,
+      expectedChecksum: "e99b23f0e723813de4f4ea0e879b46b0",
       namePrefix: "firebase-database-emulator",
     },
   },
@@ -80,15 +80,15 @@ export const DownloadDetails: { [s in DownloadableEmulators]: EmulatorDownloadDe
         },
       }
     : {
-        version: "1.10.0",
-        downloadPath: path.join(CACHE_DIR, "ui-v1.10.0.zip"),
-        unzipDir: path.join(CACHE_DIR, "ui-v1.10.0"),
-        binaryPath: path.join(CACHE_DIR, "ui-v1.10.0", "server", "server.js"),
+        version: "1.11.1",
+        downloadPath: path.join(CACHE_DIR, "ui-v1.11.1.zip"),
+        unzipDir: path.join(CACHE_DIR, "ui-v1.11.1"),
+        binaryPath: path.join(CACHE_DIR, "ui-v1.11.1", "server", "server.js"),
         opts: {
           cacheDir: CACHE_DIR,
-          remoteUrl: "https://storage.googleapis.com/firebase-preview-drop/emulator/ui-v1.10.0.zip",
-          expectedSize: 3062540,
-          expectedChecksum: "7dec1e82acccc196efc4d364e2664288",
+          remoteUrl: "https://storage.googleapis.com/firebase-preview-drop/emulator/ui-v1.11.1.zip",
+          expectedSize: 3061713,
+          expectedChecksum: "a4944414518be206280b495f526f18bf",
           namePrefix: "ui",
         },
       },
@@ -144,7 +144,13 @@ const Commands: { [s in DownloadableEmulators]: DownloadableEmulatorCommand } = 
   database: {
     binary: "java",
     args: ["-Duser.language=en", "-jar", getExecPath(Emulators.DATABASE)],
-    optionalArgs: ["port", "host", "functions_emulator_port", "functions_emulator_host"],
+    optionalArgs: [
+      "port",
+      "host",
+      "functions_emulator_port",
+      "functions_emulator_host",
+      "single_project_mode",
+    ],
     joinArgs: false,
   },
   firestore: {