+++ CHANGELOG.md
@@ -1,3 +1,3 @@
-- Enables single project mode for the auth emulator (#5068).
-- Fixes issue deploying to Hosting with i18n enabled.
-- Fixes issue where deploying to Hosting without Functions permissions would cause deployments to fail with 403 "Permission Denied" errors. (#5071)
+- Fixes issue where errors were not properly propagating when listing backends. (#5071)
+- Fixes issue where message from `-m` on deploy was not being properly applied. (#5107)
+- Fixes error `EADDRNOTAVAIL` when running emulators in Docker.